refactor: eliminate redundant wrapper components
Remove two thin wrapper components that added minimal value:

1. ControlsToolbar.tsx: Simplified wrapper around PlaybackControls
   - Removed: 68-line component file
   - Action: Inline PlaybackControls directly in SpotifyPlayerControls with TrackInfoRow/Center/Left/Right

2. VisualEffectsContainer.tsx: Wrapper around lazy-loaded VisualEffectsMenu
   - Removed: 139-line component file
   - Action: Moved lazy loading and fallback directly into PlayerContent
   - Extracted keyboard shortcuts (Ctrl+V, Ctrl+E, Ctrl+R) into new useVisualEffectsShortcuts hook

Other changes:
- Create useVisualEffectsShortcuts hook for managing visual effects keyboard shortcuts
- Fix VisualEffectsMenu onFilterChange type signature (number | boolean -> number)
- Update PlayerContent to:
  - Import and lazy-load VisualEffectsMenu directly
  - Use useVisualEffectsShortcuts hook for keyboard handling
  - Render VisualEffectsMenu conditionally with Suspense fallback

Benefits:
- Reduced component nesting (2 files removed)
- Cleaner code flow (logic moved to appropriate locations)
- Centralized keyboard shortcut handling
- All tests pass (107 tests)
- Zero type errors or lint warnings
